feat: add knife
fix: fix allowing wrong item usage

cleanup: remove some unused debug logs

+++ Assets/Scenes/Game Scene/Scripts/Effect.cs
@@ -2,10 +2,10 @@ public struct Effect
 {
     public EffectType type;
     public int magnitude;
-    public int delay;
+    public float delay;
     public int repeat;
 
-    public Effect(EffectType type, int magnitude, int delay, int repeat)
+    public Effect(EffectType type, int magnitude, float delay, int repeat)
     {
         this.type = type;
         this.magnitude = magnitude;
